Run mypy and fix typing in tests/
Previously, tests/ was excluded from mypy in the pre-commit config
because mypy was not following the configurations in setup.cfg. It turns
out mypy was not following the rules because there were some missing
__init__.py files in tests/.

This commit removes the exclusions of tests/, adds the missing
__init__.py files, updates the mypy config for tests/testing to allow
partially typed defs. Of course, these changes also exposed many typing
related issues in the test dir (and some in proxystore itself) so all of
these issues have been resolved.

+++ proxystore/p2p/manager.py
@@ -306,14 +306,14 @@ class PeerManager:
         self,
         peer_uuid: UUID,
         message: messages.PeerMessage,
-        timeout: int = 30,
+        timeout: float = 30,
     ) -> None:
         """Send message to peer.
 
         Args:
             peer_uuid (str): UUID of peer to send message to.
             message (PeerMessage): message to send to peer.
-            timeout (int): timeout to wait on peer connection to be ready.
+            timeout (float): timeout to wait on peer connection to be ready.
 
         Raises:
             PeerConnectionTimeout:
